Normalize after all, simplify, fix bugs
